requirements.txt
# === FastAPI Web Framework ===
fastapi==0.110.0
uvicorn[standard]==0.29.0

# === Core ML / LLM ===
torch>=2.0.0,<2.3.0
transformers[sentencepiece]==4.31.0
sentence-transformers==2.2.2
huggingface_hub==0.14.1
openai==1.30.1
whisper==1.1.10

python-json-logger>=2.0.7

sounddevice>=0.4.6

# === Tabular & Time Series ML (AutoGluon) ===
autogluon==1.0.0
gluonts==0.14.4

# === NLP / Embeddings ===
faiss-cpu==1.7.4
scikit-learn==1.3.2
numpy==1.26.4

# === Data Handling ===
pandas==2.1.3
python-dotenv==1.0.1
pydantic==1.10.15

# === Logging & Monitoring ===
prometheus-client==0.20.0
loguru==0.7.2

# === Explainability ===
shap==0.44.1
matplotlib==3.8.4

# === Email, HTTP, Validation ===
httpx==0.27.0
email-validator==2.1.0.post1

# === Google Calendar Support ===
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0

# === Database ===
sqlalchemy==2.0.30
psycopg2-binary==2.9.9

faster-whisper
prometheus-fastapi-instrumentator>=6.1.0
python-docx==1.1.0                # For parsing DOCX resumes
pdfminer.six==20221105            # For extracting text from PDFs
langdetect==1.0.9                 # For detecting resume language